+++ changelog.md
@@ -1,5 +1,6 @@
 # Version 2024.4.9 (2024-04-16)
 
+- Use value from service call for target (tilt) level while cover is moving
 - Decompose combined parameter
 
 # Version 2024.4.8 (2024-04-13)